refactor: Extracts signup logic from login form.
